Refactor: 예약 내역 SSR로 변경
