+++ app/(route)/my/reservations/page.tsx
@@ -1,20 +1,12 @@
-'use client';
-
-import { Suspense, useEffect, useState } from 'react';
-import { AxiosError } from 'axios';
-import { useRouter, useSearchParams } from 'next/navigation';
 import getReservations from '@/_apis/reservations/getReservations';
-import { useInfiniteQuery } from '@tanstack/react-query';
+import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
 
 import type { ReservationStatus } from '@/_types/myReservations';
-
-import { useIntersectionObserver } from '@/_hooks/activities/useIntersectionObserver';
 
 import CommonLayout from '@/_components/CommonLayout';
 import StickyLayout from '@/_components/SideStickyLayout';
 
-import FilterDropdown from './_components/FilterDropdown';
-import ReservationContainer from './_components/ReservationContainer';
+import ReservationClient from './_components/ReservationClient';
 
 export interface ReservationParams {
   cursorId?: number;
@@ -22,70 +14,27 @@ export interface ReservationParams {
   status?: ReservationStatus;
 }
 
-function MyReservations() {
-  return (
-    <Suspense fallback={<div>Loading ...</div>}>
-      <Content />
-    </Suspense>
-  );
-}
+async function MyReservations() {
+  const queryClient = new QueryClient();
 
-function Content() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-
-  const [params, setParams] = useState<ReservationParams>({
-    cursorId: undefined,
-    size: 10,
-    status: (searchParams.get('status') as ReservationStatus) || undefined,
+  await queryClient.prefetchInfiniteQuery({
+    queryKey: ['reservations', { size: 10 }],
+    queryFn: () =>
+      getReservations({
+        size: 10,
+        cursorId: undefined,
+        status: undefined,
+      }),
+    initialPageParam: undefined,
+    getNextPageParam: undefined,
   });
-
-  const { data, isError, error, fetchNextPage, hasNextPage } = useInfiniteQuery({
-    queryKey: ['reservations', params],
-    queryFn: ({ pageParam = params.cursorId }) => getReservations({ ...params, cursorId: pageParam }),
-    getNextPageParam: (lastPage) => lastPage.cursorId || undefined,
-    initialPageParam: params.cursorId,
-    retry: 0,
-  });
-
-  const showDropdown = searchParams.get('status') || data?.pages?.[0]?.totalCount !== 0;
-
-  const { setTarget } = useIntersectionObserver({
-    threshold: 0.1,
-    hasNextPage,
-    fetchNextPage,
-  });
-
-  useEffect(() => {
-    if (isError && error instanceof AxiosError) {
-      const status = error.response?.status;
-      if (status === 401) {
-        router.push('/login');
-      }
-    }
-  }, [isError, error, router]);
-
-  useEffect(() => {
-    const queryStatus = params.status ? `status=${params.status}` : '';
-    const url = `?${queryStatus}`;
-    router.replace(url);
-  }, [params, router]);
 
   return (
     <CommonLayout>
       <StickyLayout>
-        <div className="flex flex-wrap items-center justify-between gap-1">
-          <h1 className="break-keep text-3xl font-bold leading-[1.3] text-black">예약 내역</h1>
-          {showDropdown && <FilterDropdown setParams={setParams} />}
-        </div>
-        {data && (
-          <>
-            {data.pages.map((page) => (
-              <ReservationContainer key={page.cursorId ?? 'lastPage'} data={page} />
-            ))}
-            <div ref={setTarget} />
-          </>
-        )}
+        <HydrationBoundary state={dehydrate(queryClient)}>
+          <ReservationClient />
+        </HydrationBoundary>
       </StickyLayout>
     </CommonLayout>
   );
